feat(variants): SJRA-1489 extract common interpretation cell
