+++ frontend/components/stories/table/cells-mock.tsx
@@ -1,13 +1,14 @@
 import { createColumnHelper } from '@tanstack/react-table';
 
 import { Term, VepImpact } from '@/api/api';
+import GermlineInterpretationDialog from '@/apps/case/src/entity/variants/germline-occurrence/interpretation/germline-interpretation-dialog';
 import ClingenCell from '@/apps/case/src/entity/variants/germline-occurrence/table/cells/clingen-cell';
 import CNVNameCell from '@/apps/case/src/entity/variants/germline-occurrence/table/cells/cnv-name-cell';
-import HgvsgCell from '@/apps/case/src/entity/variants/table/cells/hgvsg-cell';
-import InterpretationCell from '@/apps/case/src/entity/variants/germline-occurrence/table/cells/interpretation-cell';
 import OccurrenceActionsMenu from '@/apps/case/src/entity/variants/germline-occurrence/table/cells/occurrence-actions-cell';
 import OverlapTypeGeneCell from '@/apps/case/src/entity/variants/germline-occurrence/table/cells/overlap-type-gene-cell';
 import OverlappingGeneLinkCell from '@/apps/case/src/entity/variants/germline-occurrence/table/cells/overlapping-gene-link-cell';
+import InterpretationCell from '@/apps/case/src/entity/variants/interpretation/interpretation-cell';
+import HgvsgCell from '@/apps/case/src/entity/variants/table/cells/hgvsg-cell';
 import CaseActionsMenuCell from '@/apps/case/src/exploration/table/cells/case-actions-menu-cell';
 import UninterpretedCasePreviewCell from '@/apps/variant/src/entity/cases/table/cells/uninterpreted-case-preview-cell';
 import { Status } from '@/components/base/badges/status-badge';
@@ -846,12 +847,17 @@ export const thirdSetCellData = [
 export const applicationFirstSetCellColumns = [
   baseCellColumnHelper.accessor(row => row, {
     id: 'hgvsg',
-    cell: info => <HgvsgCell occurrence={info.row.original as any} />,
+    cell: info => {
+      const row = info.row.original as any;
+      return <HgvsgCell locusId={row.locus_id} hgvsg={row.hgvsg} />;
+    },
     header: 'HgvsgCell',
   }),
   baseCellColumnHelper.accessor(row => row, {
     id: 'clinical_interpretation',
-    cell: info => <InterpretationCell occurrence={info.getValue() as any} />,
+    cell: info => (
+      <InterpretationCell occurrence={info.getValue() as any} InterpretationDialog={GermlineInterpretationDialog} />
+    ),
     header: 'InterpretationCell (Variant-Entity)',
     size: 40,
     enablePinning: false,
